Extract out expression analysis code
